limpieza

+++ Hollenhaus/Hollenhaus proyect VS/src/game/components/SendIPFromTextComponent.cpp
@@ -13,7 +13,7 @@ SendIPFromTextComponent::~SendIPFromTextComponent()
 {
 }
 
-void SendIPFromTextComponent::initComponent() 
+void SendIPFromTextComponent::initComponent()
 {
 }
 
@@ -23,5 +23,5 @@ void SendIPFromTextComponent::update()
 
 void SendIPFromTextComponent::sendIP()
 {
-	nl->SendInvitation(tc->getText().c_str(), static_cast<Uint16>(2000));
+	nl->SendInvitation(tc->getText().c_str(), 2000);
 }
